% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db}
\alias{db}
\title{Database of immune cell expression levels.}
\format{A data frame with 61 rows and 217 variables:}
\source{
\url{http://www.immgen.org}
}
\usage{
db
}
\description{
A dataset containing the expression level selected immune markers in 217 cell
types from the Immunology Genome Project (Immgen).
}
\keyword{datasets}

