% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCQ.R
\name{DCQ}
\alias{DCQ}
\title{DCQ}
\usage{
DCQ(x, db = db, alpha = 0.05, lambda.min.ratio = 0.2, nlambda = 100)
}
\arguments{
\item{x}{matrix of expression levels.}

\item{db}{databases of expression levels for markers.}

\item{alpha}{parameter passed to glmnet.}

\item{lambda.min.ratio}{parameter passed to glmnet.}

\item{nlambda}{parameter passed to glmnet.}
}
\value{
A matrix with cell abundances.
}
\description{
Perform Digital Cell Quantification
}
\examples{
NULL
}

