% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{markers}
\alias{markers}
\title{Immune markers in human and mouse.}
\format{A data frame with 61 rows and 2 variables:
\describe{
  \item{human}{immune-related markers, human name.}
  \item{mouse}{immune-related markers, mouse name.}
}}
\source{
\url{http://dcq.tau.ac.il}
}
\usage{
markers
}
\description{
A dataset containing a list of immune markers in human and mouse, used by
dcq method to infer the cell composition from RNA expression levels.
}
\keyword{datasets}
