% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCQ.R, R/dcq.R
\name{dcq}
\alias{dcq}
\alias{dcq}
\title{dcq}
\usage{
dcq(x, db = db, alpha = 0.05, lambda.min.ratio = 0.2, nlambda = 100)

dcq(x, db = db, alpha = 0.05, lambda.min.ratio = 0.2, nlambda = 100)
}
\arguments{
\item{x}{matrix of expression levels.}

\item{db}{databases of expression levels for markers.}

\item{alpha}{parameter passed to glmnet.}

\item{lambda.min.ratio}{parameter passed to glmnet.}

\item{nlambda}{parameter passed to glmnet.}

\item{x}{matrix of expression levels.}

\item{db}{databases of expression levels for markers.}

\item{alpha}{parameter passed to glmnet.}

\item{lambda.min.ratio}{parameter passed to glmnet.}

\item{nlambda}{parameter passed to glmnet.}
}
\value{
A matrix with cell abundances.

A matrix with cell abundances.
}
\description{
Perform Digital Cell Quantification

Perform Digital Cell Quantification
}
\examples{
NULL
NULL
}
